+++ webpack.prod.js
@@ -10,6 +10,11 @@ const __dirname = path.dirname(fileURLToPath(import.meta.url));
 const prod = {
   mode: "production",
   devtool: "source-map",
+  performance: {
+    hints: "warning",
+    maxAssetSize: 700 * 1024,
+    maxEntrypointSize: 800 * 1024,
+  },
   output: {
     path: path.resolve(__dirname, "dist"),
     filename: "[name].[contenthash].js",
@@ -20,7 +25,7 @@ const prod = {
     new HtmlWebpackPlugin({
       template: "./src/taskpane/taskpane.html",
       filename: "taskpane.html",
-      chunks: ["taskpane"],
+      chunks: ["runtime", "taskpane"],
       inject: "body",
       minify: {
         collapseWhitespace: true,
@@ -30,7 +35,7 @@ const prod = {
     new HtmlWebpackPlugin({
       template: "./src/commands/commands.html",
       filename: "commands.html",
-      chunks: ["commands"],
+      chunks: ["runtime", "commands"],
       inject: "body",
       minify: {
         collapseWhitespace: true,
@@ -39,6 +44,7 @@ const prod = {
     }),
   ],
   optimization: {
+    runtimeChunk: "single",
     minimize: true,
     minimizer: [
       new TerserPlugin({
@@ -51,11 +57,20 @@ const prod = {
     ],
     splitChunks: {
       chunks: "all",
+      maxInitialRequests: 12,
       cacheGroups: {
-        vendor: {
+        vendors: {
           test: /[\\/]node_modules[\\/]/,
           name: "vendors",
           chunks: "all",
+          priority: -10,
+          reuseExistingChunk: true,
+        },
+        common: {
+          minChunks: 2,
+          name: "common",
+          priority: -20,
+          reuseExistingChunk: true,
         },
       },
     },